Refactor venmic getDisplayMedia patch to share logic between preload and iframe injection
- Extract shared patch logic into venmic-patch.ts
- preload.cts now imports and calls applyVenmicPatch()
- venmic-inject.ts now imports VENMIC_PATCH_SCRIPT constant
- Add cleanup handlers (stopVenmic) in iframe injection when window.electron is available
- Document cross-origin iframe limitation where venmic cannot auto-stop
