+++ apps/desktop/src/preload.cts
@@ -106,94 +106,9 @@ contextBridge.exposeInMainWorld("electron", {
     },
 });
 
-/**
- * Patch navigator.mediaDevices.getDisplayMedia on Linux to capture audio from
- * the venmic virtual microphone (vencord-screen-share) when it's available.
- *
- * This allows screen sharing with application audio on Linux via PipeWire.
- */
+// Apply venmic patch on Linux to capture audio from virtual microphone during screen share
 if (process.platform === "linux") {
-    const originalGetDisplayMedia = navigator.mediaDevices.getDisplayMedia.bind(navigator.mediaDevices);
-
-    navigator.mediaDevices.getDisplayMedia = async function (
-        options?: DisplayMediaStreamOptions,
-    ): Promise<MediaStream> {
-        console.debug("venmic: getDisplayMedia called with options:", options);
-
-        const stream = await originalGetDisplayMedia(options);
-        console.debug(
-            "venmic: original getDisplayMedia returned stream with tracks:",
-            stream.getTracks().map((t) => `${t.kind}:${t.label}`),
-        );
-
-        // Try to find the venmic virtual microphone
-        try {
-            const devices = await navigator.mediaDevices.enumerateDevices();
-            const audioInputs = devices.filter((d) => d.kind === "audioinput");
-            console.debug(
-                "venmic: available audio input devices:",
-                audioInputs.map((d) => `${d.label} (${d.deviceId.slice(0, 8)}...)`),
-            );
-
-            const venmicDevice = devices.find((d) => d.label === "vencord-screen-share");
-
-            if (venmicDevice) {
-                console.debug("venmic: found vencord-screen-share device:", venmicDevice.deviceId);
-
-                // Capture audio from the venmic virtual microphone
-                const audioStream = await navigator.mediaDevices.getUserMedia({
-                    audio: {
-                        deviceId: { exact: venmicDevice.deviceId },
-                        autoGainControl: false,
-                        echoCancellation: false,
-                        noiseSuppression: false,
-                        channelCount: 2,
-                        sampleRate: 48000,
-                        sampleSize: 16,
-                    },
-                });
-
-                console.debug("venmic: captured audio stream from virtual mic");
-
-                // Remove any existing audio tracks and add the venmic audio
-                const audioTrack = audioStream.getAudioTracks()[0];
-                if (audioTrack) {
-                    stream.getAudioTracks().forEach((track) => {
-                        console.debug("venmic: removing existing audio track:", track.label);
-                        stream.removeTrack(track);
-                    });
-                    stream.addTrack(audioTrack);
-
-                    // Clean up venmic when the audio track ends
-                    audioTrack.addEventListener("ended", () => {
-                        console.debug("venmic: audio track ended, stopping venmic");
-                        ipcRenderer.invoke("stopVenmic").catch((e) => console.error("venmic: failed to stop:", e));
-                    });
-
-                    // Also clean up when the video track ends (screen share stopped)
-                    const videoTrack = stream.getVideoTracks()[0];
-                    if (videoTrack) {
-                        videoTrack.addEventListener("ended", () => {
-                            console.debug("venmic: video track ended, stopping venmic");
-                            ipcRenderer.invoke("stopVenmic").catch((e) => console.error("venmic: failed to stop:", e));
-                        });
-                    }
-
-                    console.log("venmic: audio track added to screen share stream");
-                } else {
-                    console.warn("venmic: no audio track returned from virtual microphone");
-                }
-            } else {
-                console.debug("venmic: vencord-screen-share device not found, audio not captured");
-            }
-        } catch (err) {
-            console.error("venmic: failed to capture audio from virtual microphone:", err);
-        }
-
-        console.debug(
-            "venmic: returning stream with tracks:",
-            stream.getTracks().map((t) => `${t.kind}:${t.label}`),
-        );
-        return stream;
-    };
+    import("./venmic-patch.js").then(({ applyVenmicPatch }) => {
+        applyVenmicPatch(() => ipcRenderer.invoke("stopVenmic"));
+    });
 }
